# src/main/resources/api-docs.yaml
openapi: 3.0.1
info:
  title: Help API
  version: "1.0"
servers:
  - url: http://192.168.25.107:8081
    description: Generated server url
paths:
...
  /menu/:
    get:
      tags:
        - menu-controller
      operationId: getProducts
      parameters:
        - name: productName
          in: query
          required: false
          schema:
            type: string
        - name: category
          in: query
          required: false
          schema:
            type: integer
            format: int32
      responses:
        "409":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "400":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "404":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "406":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "403":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
  /menu/{productCode}:
    get:
      tags:
        - menu-controller
      operationId: getProduct
      parameters:
        - name: productCode
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "409":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "400":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "404":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "406":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "403":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "200":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Product'
  /order-details/{orderId}:
    post:
      tags:
        - order-details-controller
      operationId: addOrderDetails
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/OrderWrap'
      responses:
        "409":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "400":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "404":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "406":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "403":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "200":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Orders'
  /orders/order/{orderId}:
    get:
      tags:
        - orders-controller
      operationId: getOrder
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "409":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "400":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "404":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "406":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "403":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "200":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Orders'
    post:
      tags:
        - orders-controller
      operationId: closeOrder
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "409":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "400":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "404":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "406":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "403":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "200":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Orders'
  /orders:
    post:
      tags:
        - orders-controller
      operationId: addOrder
      responses:
        "409":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "400":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "404":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "406":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "403":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "200":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Orders'
  /orders/{pageId}:
    get:
      tags:
        - orders-controller
      operationId: getOrders
      parameters:
        - name: pageId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        "409":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "400":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "404":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "406":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "403":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "200":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageOrders'
  /products:
    get:
      tags:
        - product-controller
      operationId: getProducts_2
      parameters:
        - name: productName
          in: query
          required: false
          schema:
            type: string
        - name: category
          in: query
          required: false
          schema:
            type: integer
            format: int32
        - name: active
          in: query
          required: false
          schema:
            type: boolean
            default: false
      responses:
        "409":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "400":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "404":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "406":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "403":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
    post:
      tags:
        - product-controller
      operationId: addProduct
      responses:
        "409":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "400":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "404":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "406":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "403":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "200":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Product'
  /products/product/{productCode}:
    get:
      tags:
        - product-controller
      operationId: getProduct_1
      parameters:
        - name: productCode
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "409":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "400":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "404":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "406":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "403":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "200":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Product'
    post:
      tags:
        - product-controller
      operationId: modifyProduct
      parameters:
        - name: productCode
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "409":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "400":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "404":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "406":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "403":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "200":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Product'
  /products/{pageId}:
    get:
      tags:
        - product-controller
      operationId: getProducts_1
      parameters:
        - name: pageId
          in: path
          required: true
          schema:
            type: integer
            format: int32
        - name: productName
          in: query
          required: false
          schema:
            type: string
        - name: category
          in: query
          required: false
          schema:
            type: integer
            format: int32
        - name: active
          in: query
          required: false
          schema:
            type: boolean
            default: false
      responses:
        "409":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "400":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "404":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "406":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "403":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "200":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageProduct'
...
components:
  schemas:
    ErrorDetails:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        message:
          type: string
    Category:
      type: object
      properties:
        id:
          type: integer
          format: int32
        category:
          type: string
...
    Pageable:
      type: object
      properties:
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        sort:
          $ref: '#/components/schemas/Sort'
        paged:
          type: boolean
        unpaged:
          type: boolean
        offset:
          type: integer
          format: int64
    Sort:
      type: object
      properties:
        sorted:
          type: boolean
        unsorted:
          type: boolean
        empty:
          type: boolean
    Product:
      type: object
      properties:
        id:
          type: integer
          format: int64
        productName:
          type: string
        inStockQty:
          type: number
          format: double
        activeStatus:
          type: boolean
        createdAt:
          type: string
          format: date-time
        restaurant:
          type: boolean
        unit:
          $ref: '#/components/schemas/Units'
        category:
          $ref: '#/components/schemas/Category'
    Units:
      type: object
      properties:
        id:
          type: integer
          format: int32
        unitName:
          type: string
    OrderWrap:
      type: object
      properties:
        orderId:
          type: integer
          format: int64
        productId:
          type: integer
          format: int64
        qty:
          type: number
          format: double
    OrderDetail:
      type: object
      properties:
        id:
          type: integer
          format: int64
        order:
          $ref: '#/components/schemas/Orders'
        product:
          $ref: '#/components/schemas/Product'
        quantity:
          type: number
          format: double
    Orders:
      type: object
      properties:
        orderId:
          type: integer
          format: int64
        tableId:
          type: integer
          format: int32
        waiterUsername:
          type: string
        waiterName:
          type: string
        orderStatus:
          type: string
        createdAt:
          type: string
          format: date-time
        orderDetails:
          uniqueItems: true
          type: array
          items:
            $ref: '#/components/schemas/OrderDetail'
    PageOrders:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        pageable:
          $ref: '#/components/schemas/Pageable'
        sort:
          $ref: '#/components/schemas/Sort'
        number:
          type: integer
          format: int32
        numberOfElements:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/Orders'
        empty:
          type: boolean
    PageProduct:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        pageable:
          $ref: '#/components/schemas/Pageable'
        sort:
          $ref: '#/components/schemas/Sort'
        number:
          type: integer
          format: int32
        numberOfElements:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/Product'
        empty:
          type: boolean
...